Now can approach main/kvminit

+++ kernel/types.h
@@ -5,6 +5,8 @@ typedef unsigned char  uchar;
 typedef unsigned char uint8;
 typedef unsigned short uint16;
 typedef unsigned int  uint32;
-typedef unsigned long uint64;
+// TODO: 更改了uint64
+// FIXME 
+typedef unsigned long long uint64;
 
 typedef uint64 pde_t;
